requirements.txt
Flask==3.0.3
SQLAlchemy==2.0.36
Flask-SQLAlchemy==3.1.1
Pillow==10.4.0
reportlab==4.2.5
gunicorn==21.2.0
